// types/server/contracts.d.ts
  type Vote = "yes" | "no";

  type EmbeddedWasmMsg = {
    wasm: {
      execute: {
        contract_addr: string;
        funds: Coin.Data[];
        msg: string; //base64 endocoded msg object
      };
    };
  };

  type EmbeddedBankMsg = {
    bank: {
      send: {
        amount: Coin.Data[];
        to_address: string;
      };
    };
  };
// ...
  type PageOptions = { limit?: number; start_before?: number };
  type VotesPageOptions = {
    proposal_id: number;
    limit?: number;
    start_after?: number;
  };

  type Member = {
    addr: string;
    weight: number;
  };
// ...
  type CWContracts = "apTeam" | { cw3?: string; cw4?: string };